+++ packages/react/src/types/index.ts
@@ -1,6 +1,8 @@
 export type {
   ColumnFilterType,
   IColumnFilterDef,
+  IFilterOption,
+  FilterOption,
   IColumnMeta,
   IColumnDef,
   IColumnGroupDef,